refactor sidenav location, create drawerSlice
