+++ src/features/ui/index.js
@@ -1,5 +1,5 @@
 import Layout from "features/ui/Layout";
-import SideNav from "features/ui/SideNav";
+import SideNav from "features/ui/side-nav/SideNav";
 import Header from "features/ui/Header";
 
 export { Layout, SideNav, Header };
